requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[argon2]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
